AllViewOTT_Analytics_Visualization/requirements.txt
psycopg2-binary==2.9.10
pandas==2.2.3
pymongo==4.11.1
streamlit==1.44.1
plotly==6.0.1